+++ src/js/misc.js
@@ -1,0 +1,3 @@
+global.delay = function delay(ms) {
+    return new Promise((resolve) => setTimeout(resolve, ms));
+}